pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.4",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-google-genai>=0.0.6",
    "langchain-groq>=0.1.0",
    "numpy>=1.24.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "pypdf>=3.17.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.28.0",
    "tiktoken>=0.5.0",
]
